Fixed a bug in ProjectPrefernces.removeDataSource
Added ClientPreferences.removeProjectPreference

Fixed the way I was using ClientPreferences and ProjectPreferences in
the GUI Editor. (ClientPreferences is really a wrapper for the whole
set of objects.)

Added a "remove repository" and fixed "add repository" buttons to not
throw exceptions for duplicate names (ensuring unique names).

Fixed the "add project" button to not throw exceptions for duplicate
names (ensuring unique names).