Added some guidance to the output.
Action is blank for now.
The committers are displayed in the tool tip. (cleaned up how they're displayed.)
Ease and Consequences are displayed in the tool tip (untested).
Capable affects the shading properly (untested).
When is ignored for now; it might be subsumed by Capable, the way I implemented it.
Added all the necessary icons in 32 bits.  The 16 bit system tray is displayed without guidance for now.